Completely re-built project.
The quad tree data structure is now managed by the RBQQuadTreeManager. This supports manually adding or removing items from the index and on-demand indexing of a RLMObject.

The manager saves the quad tree within a seperate RLMRealm on disk that is unique to the class name of the RLMRealm object that is being indexed.

To create a manager, you pass in an RBQIndexRequest object, which defines the entity and lat/long key paths on that entity to index.

The manager uses some simple heuristics to control when to reindex since on the fly rebalancing of the quad tree is currently not supported.

When adding more than 10k objects, the current tree structure will be deleted and completely rebuilt. This is because, rebuilding it all in memory is faster at this point than traversing the tree on disk and adding the points.

When deleting objects the data structure will be kept until the number of points drops below 80% of the original number at the last indexing.